removed approval button from comps, moved to simplified, will most likely do the same for buy/sell button
