+++ packages/comps/src/components/common/buttons.tsx
@@ -1,11 +1,7 @@
-import React, { ReactNode, useState, useCallback, useEffect } from "react";
+import React, { ReactNode } from "react";
 import Styles from "./buttons.styles.less";
 import classNames from "classnames";
 import { Spinner } from "./spinner";
-import { ApprovalAction, ApprovalState } from "../../utils/constants";
-import { AmmExchange, Cash } from "../../types";
-import { useUserStore } from "../../stores/user";
-import { approveERC20Contract } from "../../stores/use-approval-callback";
 
 export interface ButtonProps {
   id?: string;
@@ -101,114 +97,6 @@ export const TinyButton = (props: ButtonProps) => (
 export const BuySellButton = (props: ButtonProps) => (
   <Button {...props} className={classNames(Styles.BuySellButton, props.className)} />
 );
-export const ApproveButton = (props: ButtonProps) => (
-  <Button {...props} className={classNames(Styles.ApproveButton, props.className)} />
-);
-
-export const { ADD_LIQUIDITY, REMOVE_LIQUIDITY, ENTER_POSITION, EXIT_POSITION } = ApprovalAction;
-
-export const { UNKNOWN, PENDING, APPROVED } = ApprovalState;
-
-export const ApprovalButton = ({
-  amm,
-  cash,
-  actionType,
-  isApproved = false,
-  shareToken = null,
-}: {
-  amm?: AmmExchange;
-  cash: Cash;
-  actionType: ApprovalAction;
-  isApproved?: boolean;
-  shareToken?: string;
-}) => {
-  const [isPendingTx, setIsPendingTx] = useState(false);
-  const {
-    loginAccount,
-    actions: { addTransaction },
-  } = useUserStore();
-  const marketCashType = cash?.name;
-  const ammFactory = amm.ammFactoryAddress;
-  const marketDescription = `${amm?.market?.title} ${amm?.market?.description}`;
-  useEffect(() => {
-    // make sure to flip local state off if we are approved, logged, pending
-    if (isApproved && loginAccount && isPendingTx) {
-      setIsPendingTx(false);
-    }
-  }, [isApproved, loginAccount, isPendingTx]);
-
-  const approve = useCallback(async () => {
-    try {
-      setIsPendingTx(true);
-      // defaults for ADD_LIQUIDITY/most used values.
-      let ApprovalAction = approveERC20Contract;
-      let address = cash?.address;
-      let spender = ammFactory;
-      let text = `Liquidity (${marketCashType})`;
-      switch (actionType) {
-        case EXIT_POSITION: {
-          address = shareToken;
-          text = `To Sell (${marketCashType})`;
-          break;
-        }
-        case ENTER_POSITION: {
-          text = `To Buy (${marketCashType})`;
-          break;
-        }
-        case REMOVE_LIQUIDITY: {
-          address = amm?.id;
-          spender = ammFactory;
-          text = `Liquidity (${marketCashType})`;
-          break;
-        }
-        case ADD_LIQUIDITY:
-        default: {
-          break;
-        }
-      }
-      const tx = await ApprovalAction(address, text, spender, loginAccount);
-      tx.marketDescription = marketDescription;
-      addTransaction(tx);
-    } catch (error) {
-      setIsPendingTx(false);
-      console.error(error);
-    }
-  }, [cash, loginAccount, shareToken, amm]);
-
-  if (!loginAccount || isApproved) {
-    return null;
-  }
-
-  let buttonText = "";
-  let subText = "";
-  switch (actionType) {
-    case ENTER_POSITION: {
-      buttonText = "Approve to Buy";
-      break;
-    }
-    case EXIT_POSITION: {
-      buttonText = "Approve to Sell";
-      break;
-    }
-    case REMOVE_LIQUIDITY: {
-      buttonText = "Approve Removal";
-      subText = "(approve to see removal estimation)";
-      break;
-    }
-    default:
-      buttonText = `Approve ${marketCashType}`;
-      break;
-  }
-
-  return (
-    <ApproveButton
-      disabled={isPendingTx}
-      text={isPendingTx ? "Approving..." : buttonText}
-      subText={subText}
-      action={() => approve()}
-    />
-  );
-};
 
 export enum BUTTON_TYPES {
   PRIMARY = "Primary",
@@ -222,6 +110,7 @@ export interface BaseThemeButtonProps {
   action?: Function;
   customContent?: any;
   customClass?: any;
+  customClassName?: string;
   subText?: string;
   icon?: any;
   disabled?: boolean;
