+++ packages/simplified/src/modules/modal/modal-add-liquidity.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useMemo, useState } from "react";
 import Styles from "./modal.styles.less";
 import { Header } from "./common";
 import { useHistory } from "react-router";
-import { InfoNumbers } from "../market/trading-form";
+import { InfoNumbers, ApprovalButton } from "../market/trading-form";
 import classNames from "classnames";
 import { AmmOutcome, Cash, LiquidityBreakdown, MarketInfo, DataState } from "@augurproject/comps/build/types";
 import { BigNumber as BN } from "bignumber.js";
@@ -29,7 +29,7 @@ const {
 const { formatPercent, formatSimpleShares, formatEther } = Formatter;
 const {
   Icons: { BackIcon },
-  ButtonComps: { ApprovalButton, BuySellButton },
+  ButtonComps: { BuySellButton },
   SelectionComps: { MultiButtonSelection },
   InputComps: { AmountInput, isInvalidNumber, OutcomesGrid },
   LabelComps: { generateTooltip, WarningBanner },
